==> tag-classificaton/1.1 DataPreparation.ipynb <==
!pip3 install openpyxl

# import json 
import pandas as pd 
from pathlib import Path 
from glob import glob
import os

BASE_DIR = "./data/raw_input/" 

dfs = [] 
ignored_files = []
for filepath in glob(BASE_DIR + "/*"): 
    try: 
        tempdf = pd.read_json(filepath) 
        filename = Path(filepath).name
        tempdf['filename'] = filename
        dfs.append(tempdf) 
    except Exception as e: 
        ignored_files.append(filepath)

df = pd.concat(dfs, ignore_index=False).reset_index(drop=True)
df = df[df.tags.map(lambda x: len(x)) <= 1]
df['tags'] = df.tags.map(lambda x: x[0].strip() if x else "NA")

print(f"Number of rows: {df.shape[0]}")
print(f"Number of unique files: {pd.unique(df.filename).shape[0]}")
print(f"Number of unique tags: {pd.unique(df.tags).shape[0]}")

all_tags = set(df.fillna("NA").tags.unique())

df[['tags']].groupby('tags').size().reset_index()

## Load Tag Mapping 

xl = pd.ExcelFile("./data/Updated-Tagging.xlsx")

xl.sheet_names

tagging_df = pd.read_excel("./data/Updated-Tagging.xlsx", sheet_name = "AI Tag List")

standard_tags = set(tagging_df['Standard Tags'].unique())
non_standard_tags = set(tagging_df['Tusk Compare Fields'].unique())

nst_to_st = {t['Tusk Compare Fields'].strip() : t['Standard Tags'].strip() for t in tagging_df.to_dict(orient="records")}
for t in tagging_df.to_dict(orient="records"): 
    nst_to_st[t['Standard Tags'].strip()] = t['Standard Tags'].strip()

st_to_nst = {} 
for key, value in nst_to_st.items(): 
    st_to_nst[value] = key 

visited_st = {key: False for key in st_to_nst}
len(visited_st)

# print("Tags missing from the Updated Tagging:")
# for i, t in enumerate(all_tags - (standard_tags.union(non_standard_tags)), start=1):
#     print(i, t)
# print()
# print("Tags Missing from the data:")
# for i, t in enumerate((standard_tags.union(non_standard_tags)) - all_tags, start=1):
#     print(i, t, " - ",  nst_to_st[t])

tagging_df[['Standard Tags']].groupby("Standard Tags").size().reset_index()

nst_to_st.update({
    "Facilities / Instruments": "Facilities / Instrument", 
    "Facilities": "Incremental Facilities", 
    "": "NA", 
    "Governing Law": "Governing Laws", 
    "Default" : "Event of Default",
    "Event of Default": "Event of Default",
    "Facilities" : "Incremental Facilities",
    "ESG Reporting Obligations": "Reporting Requirements", 
    "Subsidiary" : "NA",
    "Collateral" : "NA",
    "Subordination" : "NA",
    "Defaulted Interest" : "Pricing / Interest", 
    "Prepayments" : "Prepayment / Redemption",
    "Interest Payment Dates":  "Pricing / Interest",
    "NA": "NA",
    "Use of Proceeds": "NA", 
    "Other Reporting Requirements": "Reporting Requirements", 
    "Interest / Pricing": "Pricing / Interest",

    
})

for tag in all_tags: # all tags from dataset 
    if tag in nst_to_st:
        visited_st[nst_to_st[tag]] = True
    else:
        print(tag)

for key, visited in visited_st.items(): 
    if not visited: 
        print(key)

print("Total Size of Updated Tagging", tagging_df.shape)

df[['tags']].groupby('tags').size().reset_index().sort_values(by=0, ascending=False)

print(f"Number of files with empty section content: {pd.unique(df[df['section_content'].isna()].filename).shape[0]}")
print(f"Number of files with empty section and tag present content: {pd.unique(df[(df['section_content'].isna()) & (df['tags'].notna()) ].filename).shape[0]}")
print(f"Number of records where section is empty: {df[(df['section_content'].isna())].shape[0]}")

df['section_content'] = df['section_content'].fillna("")
df['tags'] = df['tags'].fillna("NA")
df['title'] = df['title'].fillna("") 
df['category'] = df['category'].fillna("")

df['word_count'] = df['section_content'].fillna("").apply(lambda x : len(x.split()))

# Filter section content with small text

df = df[(df.word_count > 20)]
df = df.fillna("NA") 

# df['section_content'] = df['title'].apply(lambda x: f"Title is {x}") + "\n" + df['category'].apply(lambda x: f"Category is {x}") + "\n" +  df['section_content']
df['section_content'] = df['title'].apply(lambda x: f"Title is {x}") + "\n" +  df['section_content']

df['standard_tags'] = df.tags.map(lambda x: nst_to_st[x])

df.to_json("./data/all-data-raw-standard-tags.json")

df.word_count.describe()

# Preprocessing 

df = pd.read_json("./data/all-data-raw-standard-tags.json")

df = df.drop_duplicates('section_content') 

## Select content only form shortlisted files

file_analysis_df = pd.read_excel("./data/File-Analysis.xlsx")

good_files = set(file_analysis_df.File.to_list()[:100])

df['tags'] = df['tags'].str.replace("Events of Default", "Event of Default")

# df = df[df.filename.str.contains("Agreement", case=False)]

df = df.drop_duplicates('section_content', keep=False)

df = df[df.filename.map(lambda x: x in good_files)]

df['standard_tags'] = df.tags.map(lambda x: nst_to_st[x])

df[['standard_tags']].groupby('standard_tags').size().reset_index().sort_values(by=0)

df[['tags']].groupby('tags').size().reset_index().sort_values(by=0)

df.word_count.describe()

df =  df.drop(columns = 'tags')

df = df.rename(columns = {'standard_tags': 'tags'})
df.head()

print(df.shape)

for t in df[['tags']].groupby('tags').size().reset_index().sort_values(by=0).itertuples(): 
    print(t.tags,":", t._2)

df.word_count.describe()

df[df.tags == "NA"][['tags', 'title', 'category']].groupby('tags').agg(set)

agg_df = df[['tags', 'title' ]].groupby([ 'tags', 'title']).size().reset_index().sort_values(by=0).rename(columns={0:'freq'})

tag_title_filter = {t.tags+""+t.title for t in agg_df[agg_df.tags == "NA"][agg_df.freq <=3].itertuples()}

len(tag_title_filter)

df['tag_title'] = df.tags+df.title

new_df = pd.concat([df[df.tags!="NA"], df[df.tag_title.isin( tag_title_filter)]])

new_df.word_count.describe()

sum(new_df[['tags']].groupby('tags').size().reset_index().sort_values(by=0).rename(columns={0:"freq"}).freq.values)

# Save Data cleaned for NA tags

new_df.to_json("./data/cleaned_tags_data_na_tag_v2_lt3.json", orient='records')

print(new_df.shape)

from sklearn.model_selection import KFold
from sklearn.model_selection import train_test_split

unique_filenames = df.filename.unique()

# TODO  Check correct data frame for train test split

# Split filenames into train and test sets
train_files, test_files = train_test_split(unique_filenames, test_size=0.2, random_state=42)

# Create train and test datasets based on the split filenames
train_df = new_df[new_df['filename'].isin(train_files)].reset_index() 
test_df = df[df['filename'].isin(test_files)].reset_index()
del train_df['index'] 
del test_df['index'] 
train_df.shape, test_df.shape

selected_columns = ["section_content", "tags"]

train_tags = set(train_df.tags.unique())
test_tags = set(test_df.tags.unique())

print(test_tags - train_tags )
assert not test_tags - train_tags 

OUTPUT_DIR = "./data/dataset/v8"
os.makedirs(OUTPUT_DIR, exist_ok=True)

train_df.head()

print(train_df.shape)
print(test_df.shape)

train_df.to_json(os.path.join(OUTPUT_DIR, "train.json")) 
test_df.to_json(os.path.join(OUTPUT_DIR, "test.json"))

# Integrate Domain Correction 

import pandas as pd

xl = pd.ExcelFile("./data/Predictions_v13_small_with_sec_numbering_cleaned.xlsx")

xl.sheet_names

updated_df = pd.read_excel("./data/Predictions_v13_small_with_sec_numbering_cleaned.xlsx")

updated_df.columns

updated_df = updated_df[['filename', 'title', 'section_number', 'section_content',
       'Original_Tag', 'Predicted_Tag', 'match', 'Domain']]

updated_df = updated_df.fillna("NA")

updated_df['Domain'] = updated_df.Domain.str.strip()

domain_tags = set(updated_df.Domain.values)
all_tags = set(updated_df.Original_Tag.values)
print(len(domain_tags))
print(len(all_tags))
print("\nCorrection Tags:")
print("\n".join(domain_tags - all_tags))

print("\n".join(sorted(all_tags)))

nst_to_st = {
    "Interest / Pricing": "Pricing / Interest",
    "Other Reporting Requirements": "Reporting Requirements"
}

def to_standard_tag(x): 
    if x in nst_to_st: 
        return nst_to_st.get(x)
    return x
        
    

updated_df['Domain'] = updated_df.Domain.apply(to_standard_tag)

updated_df = updated_df[updated_df.Domain.map(lambda x: x in all_tags)]

domain_tags = set(updated_df.Domain.values)
all_tags = set(updated_df.Original_Tag.values)
print(len(domain_tags))
print(len(all_tags))
print("\nCorrection Tags:")
print("\n".join(domain_tags - all_tags))

print("\n".join(sorted(domain_tags)))

updated_df['tags'] = updated_df['Domain']

# updated_df.to_json("./data/dataset/v9/train.json")
updated_df.to_excel("./data/v9_train.xlsx", index=False)

"""
1. Domain NA Tags 
2. Multiple tags 
3. Blank means ok. 
Optional Prepayment -> Optional Prepayment / Redemption
"""
