==> tag-classificaton/2.1 TagClassificationV3-KV.ipynb <==
# Model Training 

!pip install -q transformers[torch] accelerate
!pip3 install -q wandb
!pip3 install -q openpyxl

import json 
import pandas as pd 
import os 
import pickle

from sklearn.metrics import accuracy_score, precision_recall_fscore_support
import numpy as np 

from transformers import BertTokenizer, BertForSequenceClassification
from transformers import DistilBertTokenizer, DistilBertForSequenceClassification
from transformers import Trainer, TrainingArguments
import torch
from torch.utils.data import Dataset
from sklearn.metrics import classification_report, precision_recall_curve
from sklearn.preprocessing import LabelEncoder


# !pip3 uninstall -q wandb -y

import wandb

wandb.login(key="4de6103347df6561e7258cdef0ef60bbc1233695", relogin=True)

MODEL_DIRECTORY = "./trained_model/v15-small"
DATA_DIRECTORY = "./data/dataset/v9/"

class TextDataset(Dataset):
    def __init__(self, texts, labels, tokenizer, max_length):
        self.texts = texts
        self.labels = labels
        self.tokenizer = tokenizer
        self.max_length = max_length

    def __len__(self):
        return len(self.texts)

    def __getitem__(self, idx):
        text = self.texts[idx]
        labels = self.labels[idx]
        encoding = self.tokenizer.encode_plus(
            text,
            max_length=self.max_length,
            truncation=True,
            padding='max_length',
            return_tensors='pt',
            return_attention_mask=True
        )
        return {
            'input_ids': encoding['input_ids'].flatten(),
            'attention_mask': encoding['attention_mask'].flatten(),
            'labels': torch.tensor(labels, dtype=torch.long)
        }

train_df = pd.read_json(os.path.join(DATA_DIRECTORY, "train.json"))

print(train_df.shape)

print(train_df.filename.unique().shape)

texts = train_df['section_content'].values
labels_text = train_df['tags'].values

label_encoder = LabelEncoder()
labels = label_encoder.fit_transform(labels_text)

os.makedirs(MODEL_DIRECTORY)

with open(os.path.join(MODEL_DIRECTORY, "tag.le"), 'wb') as file: 
    pickle.dump(label_encoder, file)

num_labels = train_df.tags.unique().shape[0]
print("Number of labels", num_labels)

## Calculate Class weights

from sklearn.utils.class_weight import compute_class_weight
import numpy as np

class_weights = compute_class_weight(class_weight='balanced', classes=np.unique(labels), y=labels)
class_weights_dict = {i: weight for i, weight in enumerate(class_weights)}

print("Class Weights: ", class_weights_dict)


class_weight_tensor = torch.tensor(class_weights, dtype=torch.float) 

def compute_weighted_loss(model, inputs, return_outputs=False): 
    labels = inputs.get("labels") 
    outputs = model(**inputs)
    logits = outputs.get("logits")
    loss_fct = torch.nn.CrossEntropyLoss(weight=class_weight_tensor.to(logits.device))
    loss = loss_fct(logits, labels) 
    return (loss, outputs) if return_outputs else loss

# model_name = 'bert-base-uncased'
# tokenizer = BertTokenizer.from_pretrained(model_name)
# model = BertForSequenceClassification.from_pretrained(model_name, num_labels=num_labels)

## Load Model

tokenizer = DistilBertTokenizer.from_pretrained('distilbert-base-uncased')
# model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased', num_labels=num_labels)

dataset = TextDataset(texts, labels, tokenizer, max_length=512)

## Compute Metrics callback

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    
    accuracy = accuracy_score(labels, predictions, )
    precision = precision_score(labels, predictions, average='macro')
    recall = recall_score(labels, predictions, average='macro')
    f1 = f1_score(labels, predictions, average='macro')
    
    return {
        "accuracy": accuracy,
        "precision": precision,
        "recall": recall,
        "f1": f1
    }

MODEL_DIRECTORY

class CustomTrainer(Trainer):
    def __init__(self, *args, class_weights=None, **kwargs):
        super().__init__(*args, **kwargs)
        self.class_weights = class_weights

    def compute_loss(self, model, inputs, return_outputs=False):
        labels = inputs.get("labels")
        outputs = model(**inputs)
        logits = outputs.get("logits")
        loss_fct = torch.nn.CrossEntropyLoss(weight=self.class_weights.to(logits.device))
        loss = loss_fct(logits, labels)
        return (loss, outputs) if return_outputs else loss

## Train using K-fold Cross Validation

from sklearn.model_selection import StratifiedKFold, KFold
import json 
from pathlib import Path

k=5
kf = KFold(n_splits=k, shuffle=True, random_state=42)

unique_files = train_df[train_df.filename.str.contains("synth") == False].filename.unique()

splits = list(kf.split(unique_files))
for fold, (train_file_index, val_file_index)  in enumerate(splits):
    train_file_names = set(unique_files[train_file_index])
    val_file_names = set(unique_files[val_file_index])
    
    train = train_df[train_df.filename.isin(train_file_names)]
    val  = train_df[train_df.filename.isin(val_file_names)]
    print(train.shape, train.tags.unique().shape)
    print(val.shape, val.tags.unique().shape)

## Split by files

unique_labels = train_df.tags.unique()

fold_accuracies = [] 
for fold, (train_file_index, val_file_index)  in enumerate(splits):
    
    train_file_names = set(unique_files[train_file_index])
    val_file_names = set(unique_files[val_file_index])
    
    train = train_df[train_df.filename.isin(train_file_names)]
    val  = train_df[train_df.filename.isin(val_file_names)]
    
    val_labels = label_encoder.transform(val['tags'].values)
    train_labels = label_encoder.transform(train['tags'].values)
    
    train_dataset = TextDataset(train['section_content'].values, train_labels ,  tokenizer, max_length=512)
    val_dataset = TextDataset(val['section_content'].values, val_labels,  tokenizer, max_length=512)
    
    model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased', num_labels=num_labels)
    
    run_name = f"{Path(MODEL_DIRECTORY).name}_fold_{fold+1}"    
    print("Starting ", run_name)
    wandb.init(project="huggingface", name=run_name)
    
    training_args = TrainingArguments(
        output_dir=os.path.join(MODEL_DIRECTORY, f'./results_fold_{fold+1}'),
        run_name=run_name,
        num_train_epochs=30,
        per_device_train_batch_size=16,
        per_device_eval_batch_size=16,
        weight_decay=0.01,
        logging_dir=os.path.join(MODEL_DIRECTORY, f'./logs_fold_{fold+1}'),
        logging_steps=10,
        learning_rate=2e-5,
        eval_strategy="epoch", 
        save_strategy="epoch",
        load_best_model_at_end=True,
        metric_for_best_model='recall'

    )

    # Initialize Trainer
    trainer = CustomTrainer(
        model=model,
        args=training_args,
        train_dataset=train_dataset,
        eval_dataset=val_dataset, 
        compute_metrics=compute_metrics, 
        class_weights=class_weight_tensor, 
    )
    trainer.train()

    # Train the model
    # Make predictions on the test set
    predictions = trainer.predict(val_dataset)

    # Extract the logits and convert to predicted labels
    logits = predictions.predictions
    predicted_labels = np.argmax(logits, axis=-1)

    # Convert numeric labels back to original text labels for comparison
    predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
    true_labels_text = label_encoder.inverse_transform(val_labels)
    
    val['predicted_tags'] = predicted_labels_text
    val.to_json(f"./data/output/val_{Path(MODEL_DIRECTORY).name}_fold_{fold+1}.json", orient="records")
    # Generate a classification report
    report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_, labels=unique_labels, output_dict=True)
    with open(f"./data/output/tag_classification_fold_{fold+1}_report.json", "w") as file: 
        json.dump(report, file)
        
    print(report)
        
    
    print("Generating Evaluation results")
    eval_result = trainer.evaluate()
    # Save accuracy for this fold
    fold_accuracies.append((report, eval_result['eval_accuracy']))
    del model 
    torch.cuda.empty_cache()

---

## Split by rows

fold_accuracies = [] 
for fold, (train_index, val_index) in enumerate(kf.split(train_df, train_df['tags'].values)):
    print(f"Fold {fold + 1}/{k}")
    train = train_df.iloc[train_index]
    val = train_df.iloc[val_index]
    
    val_labels = label_encoder.transform(val['tags'].values)
    train_labels = label_encoder.transform(train['tags'].values)
    
    train_dataset = TextDataset(train['section_content'].values, train_labels ,  tokenizer, max_length=512)
    val_dataset = TextDataset(val['section_content'].values, val_labels,  tokenizer, max_length=512)
    
    model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased', num_labels=num_labels)
    
    run_name = f"{Path(MODEL_DIRECTORY).name}_fold_{fold+1}"
    wandb.run_name = run_name


    
    print("Starting ", run_name)
    
    training_args = TrainingArguments(
        output_dir=os.path.join(MODEL_DIRECTORY, f'./results_fold_{fold+1}'),
        run_name=run_name,
        num_train_epochs=30,
        per_device_train_batch_size=16,
        per_device_eval_batch_size=16,
        weight_decay=0.01,
        logging_dir=os.path.join(MODEL_DIRECTORY, f'./logs_fold_{fold+1}'),
        logging_steps=10,
        learning_rate=2e-5,
        eval_strategy="epoch", 
        save_strategy="epoch",
        load_best_model_at_end=True,
        metric_for_best_model='recall'

    )

    # Initialize Trainer
    trainer = CustomTrainer(
        model=model,
        args=training_args,
        train_dataset=train_dataset,
        eval_dataset=val_dataset, 
        compute_metrics=compute_metrics, 
        class_weights=class_weight_tensor, 
    )
    trainer.train()

    # Train the model
    # Make predictions on the test set
    predictions = trainer.predict(val_dataset)

    # Extract the logits and convert to predicted labels
    logits = predictions.predictions
    predicted_labels = np.argmax(logits, axis=-1)

    # Convert numeric labels back to original text labels for comparison
    predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
    true_labels_text = label_encoder.inverse_transform(val_labels)
    
    val['predicted_tags'] = predicted_labels_text
    val.to_json(f"./data/output/val_{Path(MODEL_DIRECTORY).name}_fold_{fold+1}.json", orient="records")
    # Generate a classification report
    report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_, output_dict=True)
    with open(f"./data/output/tag_classification_fold_{fold+1}_report.json", "w") as file: 
        json.dump(report, file)
        
    print(report)
        
    
    print("Generating Evaluation results")
    eval_result = trainer.evaluate()
    # Save accuracy for this fold
    fold_accuracies.append((report, eval_result['eval_accuracy']))
    del model 
    torch.cuda.empty_cache()

print("")

x[0]

fold_data = []
for x in fold_accuracies[1:]:
    for key, value in x[0].items(): 
        if type(value) != dict:
            continue
        fold_data.append({
            "tag": key, 
            "precision": value['precision'], 
            'recall' : value['recall'], 
            'f1-score': value['f1-score'], 
            'support': value['support']
        })
    break

for i in fold_accuracies: 
    print(i[1])

fold_df = pd.DataFrame(fold_data)

print(fold_df.groupby('tag').agg('mean').to_string())

## Normal Training 

model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased', num_labels=num_labels)

dataset = TextDataset(texts, labels, tokenizer, max_length=512)

import os 
from pathlib import Path


training_args = TrainingArguments(
    output_dir=os.path.join(MODEL_DIRECTORY, './results'),
    run_name=f"{Path(MODEL_DIRECTORY).name}",
    num_train_epochs=50,
    per_device_train_batch_size=16,
    per_device_eval_batch_size=16,
    weight_decay=0.01,
    logging_dir=os.path.join(MODEL_DIRECTORY, './logs'),
    logging_steps=10,
    learning_rate=2e-5,
    eval_strategy="epoch"
)

# Initialize Trainer
trainer = CustomTrainer(
    model=model,
    args=training_args,
    train_dataset=dataset,
    eval_dataset=dataset, 
    compute_metrics=compute_metrics, 
    class_weights=class_weight_tensor
)

# Train the model


trainer.train()

# eval_results = trainer.evaluate()
# print(f"Evaluation results: {eval_results}")

# Save the model and tokenizer to disk
model.save_pretrained(MODEL_DIRECTORY)
tokenizer.save_pretrained(MODEL_DIRECTORY)

# Load Model For Inferences

!pip install -q transformers[torch] accelerate
!pip3 install -q wandb
!pip3 install -q openpyxl

MODEL_DIRECTORY = "./trained_model/v15-small"
DATA_DIRECTORY = "./data/carlyle-inference"

import json 
import pandas as pd 
import os 
import pickle
from glob import glob

from sklearn.metrics import accuracy_score, precision_recall_fscore_support
import numpy as np 
from sklearn.metrics import classification_report

from transformers import BertTokenizer, BertForSequenceClassification
from transformers import DistilBertTokenizer, DistilBertForSequenceClassification
from transformers import Trainer, TrainingArguments
import torch
from torch.utils.data import Dataset
from sklearn.metrics import classification_report, precision_recall_curve
from sklearn.preprocessing import LabelEncoder


## TextDatset Class

class TextDataset(Dataset):
    def __init__(self, texts, labels, tokenizer, max_length):
        self.texts = texts
        self.labels = labels
        self.tokenizer = tokenizer
        self.max_length = max_length

    def __len__(self):
        return len(self.texts)

    def __getitem__(self, idx):
        text = self.texts[idx]
        labels = self.labels[idx]
        encoding = self.tokenizer.encode_plus(
            text,
            max_length=self.max_length,
            truncation=True,
            padding='max_length',
            return_tensors='pt',
            return_attention_mask=True
        )
        return {
            'input_ids': encoding['input_ids'].flatten(),
            'attention_mask': encoding['attention_mask'].flatten(),
            'labels': torch.tensor(labels, dtype=torch.long)
        }

## Inference 

print(MODEL_DIRECTORY)

model = DistilBertForSequenceClassification.from_pretrained(MODEL_DIRECTORY)
tokenizer = DistilBertTokenizer.from_pretrained(MODEL_DIRECTORY)

with open(os.path.join(MODEL_DIRECTORY, "tag.le"), 'rb') as file: 
    label_encoder = pickle.load(file)

inputs = "This is a sample text"

tokens = tokenizer(inputs, return_tensors='pt', padding=True, truncation=True)

data = tokens
with torch.no_grad():
    outputs = model(**data)

logits = outputs.logits
predictions = torch.argmax(logits, dim=-1).tolist()

predictions

## Inference Trainer

inference_args = TrainingArguments(
    output_dir='./results',  
    per_device_eval_batch_size=32, 
    dataloader_drop_last=False,  
    no_cuda=False if torch.cuda.is_available() else True, 
    seed=42,  
    report_to="none" # Disable wandb reporting. 
)
trainer = Trainer(
    model=model, 
    args = inference_args
)

import re
def extract_section_num(x): 
    text = x.split("\n")[1].strip().upper()
    matches = re.findall("^SECTION\s+\d+\.\d{1,2}", text)
    if matches:
        return matches[0]
    matches  = re.findall("^SECTION\s+\d+\.?\d{1,2}", text)
    if matches:
        return matches[0]
    matches  = re.findall("^\d+\.?\d{1,2}", text, flags=re.IGNORECASE)
    if matches:
        return matches[0]
    return "NOT FOUND"

def run_predict(filename): 
    df = pd.read_json(filename)
    df = df[df.section_content.isna() == False]
    df['section_content'] = df['title'].apply(lambda x: f"Title is {x}") + "\n" +  df['section_content']
    test_texts = df['section_content'].values
    test_dataset = TextDataset(test_texts, [0] * len(test_texts), tokenizer, max_length=512)
    predictions = trainer.predict(test_dataset)

    # Extract the logits and convert to predicted labels
    logits = predictions.predictions
    predicted_labels = np.argmax(logits, axis=-1)

    # Convert numeric labels back to original text labels for comparison
    predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
    df['toc'] = df.section_content.apply(extract_section_num)

    df['AITag'] = predicted_labels_text
    df['AITag'] = df.AITag.str.replace("^NA$", "", regex=True)
    df["ExpertTag"] = df['AITag']
    df = df[['toc', 'title', 'section_content', 'AITag', 'ExpertTag']]
    return df 

# filename = os.path.join(DATA_DIRECTORY, "test.json")

filenames = glob(os.path.join(DATA_DIRECTORY, "*/*.section.json"))
filename = filenames[0]
cat_filenames = glob(os.path.join(DATA_DIRECTORY, "*/categorized.xlsx"))

from pathlib import Path
import os 
import shutil

def extract_num(x): 
    try: 
        return "".join(re.findall("\d", x))
    except Exception as e: 
        print(x)
        raise e

for filename, cat_filename in zip(filenames, cat_filenames): 
    test_df = run_predict(filename)
    cat_df = pd.read_excel(cat_filename)
    cat_df = cat_df.drop(columns=["AITag", "ExpertTag"])
    
    output_dir = os.path.join(DATA_DIRECTORY, Path(filename).stem.split(".")[0])
    output_file = os.path.join(output_dir, Path(filename).stem.split(".")[0]+".tags.xlsx")
    root_output_file = os.path.join(Path(output_file).parent.parent, Path(filename).stem.split(".")[0]+".tags.xlsx")
    
    # Create keys for joining
    test_df['key'] = test_df.toc.apply(extract_num) + "_" + test_df.title.str.lower()
    cat_df['key'] = cat_df.toc_number.map(str).apply(extract_num) + "_" + cat_df.title.str.lower()
    
    result_df = cat_df.merge(test_df[['AITag', 'ExpertTag', 'key']], on='key', how='left').fillna("   ")
    result_df = result_df.drop(columns = "key")
    print(output_dir)
    print(output_file)
    print(result_df.shape, result_df[result_df.AITag.isna()].shape)
    result_df.to_excel(output_file, index=False)
    result_df.to_excel(root_output_file, index=False)
    

output_file

cat_df.head()

test_df.head()

cat_df

test_df[test_df.AITag!=""]

print(output_file)

test_df.head()

# test_labels = test_df['tags'].values
# test_labels = label_encoder.transform(test_labels)

# Make predictions on the test set
predictions = trainer.predict(test_dataset)

# Extract the logits and convert to predicted labels
logits = predictions.predictions
predicted_labels = np.argmax(logits, axis=-1)

# Convert numeric labels back to original text labels for comparison
predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
# true_labels_text = label_encoder.inverse_transform(test_labels)

# # Generate a classification report
# report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_)


---

nst_to_st = {
    "Interest / Pricing": "Pricing / Interest",
    "Other Reporting Requirements": "Reporting Requirements"
}

def to_standard_tag(x): 
    if x in nst_to_st: 
        return nst_to_st.get(x)
    return x
        
    

test_df['tags'] = test_df.tags.apply(to_standard_tag)

test_texts = test_df['section_content'].values
test_labels = test_df['tags'].values
test_labels = label_encoder.transform(test_labels)

from sklearn.metrics import classification_report

test_dataset = TextDataset(test_texts, test_labels, tokenizer, max_length=512)

inference_args = TrainingArguments(
    output_dir='./results',  
    per_device_eval_batch_size=32, 
    dataloader_drop_last=False,  
    no_cuda=False if torch.cuda.is_available() else True, 
    seed=42,  
    report_to="none" # Disable wandb reporting. 
)
trainer = Trainer(
    model=model, 
    args = inference_args
)

os.environ["WANDB_DISABLED"] = "true"

## Training Accuracy

# Make predictions on the test set
predictions = trainer.predict(dataset)

# Extract the logits and convert to predicted labels
logits = predictions.predictions
predicted_labels = np.argmax(logits, axis=-1)

# Convert numeric labels back to original text labels for comparison
predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
true_labels_text = label_encoder.inverse_transform(labels)

# Generate a classification report
report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_)


print(report)

## Test Accuracy.

# Make predictions on the test set
predictions = trainer.predict(test_dataset)

# Extract the logits and convert to predicted labels
logits = predictions.predictions
predicted_labels = np.argmax(logits, axis=-1)

# Convert numeric labels back to original text labels for comparison
predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
true_labels_text = label_encoder.inverse_transform(test_labels)

# Generate a classification report
report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_)


print(report)

test_df['Predicted_Tag'] = predicted_labels_text

test_df = test_df.rename(columns = {"tag": "Original_Tag"})

print(test_df.shape)

test_df.to_excel("./data/output/predictions_v15_small.xlsx")

## Generate Results on Unsampled NA data

seen_df = pd.read_json("./data/cleaned_tags_data.json")
all_df = pd.read_json("./data/all-data.json")

unseen_df = all_df.merge(seen_df, how='left', on=['filename', 'tags', 'title', 'category', 'word_count', 'section_content'], indicator=True, )
unseen_df = unseen_df[unseen_df['_merge'] == 'left_only']
unseen_df = unseen_df.drop(columns=['_merge'])



unseen_df['tags'] = unseen_df['tags'].str.replace("Events of Default", "Event of Default")

unseen_texts = unseen_df['section_content'].values
unseen_labels = unseen_df['tags'].values
unseen_labels = label_encoder.transform(unseen_labels)

unseen_dataset = TextDataset(unseen_texts, unseen_labels, tokenizer, max_length=512)

# Make predictions on the test set
predictions = trainer.predict(unseen_dataset)

# Extract the logits and convert to predicted labels
logits = predictions.predictions
predicted_labels = np.argmax(logits, axis=-1)

# Convert numeric labels back to original text labels for comparison
predicted_labels_text = label_encoder.inverse_transform(predicted_labels)
true_labels_text = label_encoder.inverse_transform(unseen_labels)

# Generate a classification report
report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_)


print(report)

all_df = pd.read_json("./data/all-data.json")
print(all_df.shape)
all_df = all_df[all_df.word_count > 20]
print(all_df.shape)

test_df.head()

test_df.shape

test_df['predicted_tags'] = predicted_labels_text

test_df.head()

group_df = all_df[['section_content', 'tags']].groupby('section_content').agg(list).reset_index()
group_df['n_tags'] = group_df.tags.apply(lambda x: len(x))
group_df['n_unique_tags'] = group_df.tags.apply(lambda x: len(set(x)))
print(group_df[group_df.n_tags > 1].shape)
print(group_df[group_df.n_unique_tags > 1].shape)
group_df.sort_values(by='n_unique_tags', ascending=False).head(10)

test_df[["section_content", "predicted_tags"]].merge(all_df.drop_duplicates("section_content"), how="left", on="section_content", validate="one_to_one").to_excel("./data/output/test_data_predictions.xlsx")

test_df.shape

print(report)

def predict_labels(texts, model, tokenizer, label_encoder, max_length=512, device='cuda'):
    # Move model to the specified device
    model.to(device)
    
    # Tokenize the texts
    encodings = tokenizer(
        texts,
        max_length=max_length,
        truncation=True,
        padding=True,
        return_tensors="pt"
    )
    
    # Move tensors to the same device as the model
    input_ids = encodings['input_ids'].to(device)
    attention_mask = encodings['attention_mask'].to(device)
    
    # Put the model in evaluation mode
    model.eval()
    
    # Perform the inference
    with torch.no_grad():
        outputs = model(
            input_ids=input_ids,
            attention_mask=attention_mask
        )
        logits = outputs.logits
        predictions = torch.argmax(logits, dim=-1)
    
    # Move predictions back to the CPU before converting to numpy
    predicted_labels = label_encoder.inverse_transform(predictions.cpu().numpy())
    
    return predicted_labels

predicted_labels_text = predict_labels(test_texts.tolist(), model, tokenizer, label_encoder)
true_labels_text = label_encoder.inverse_transform(test_labels)

!nvidia-smi

# Extract the logits and convert to predicted labels
# logits = predictions.predictions
# predicted_labels = np.argmax(logits, axis=-1)

# Convert numeric labels back to original text labels for comparison
# Generate a classification report
report = classification_report(true_labels_text, predicted_labels_text, target_names=label_encoder.classes_)


print(report)
