==> tag-classificaton/3.1 CrossValidationEvaluation.ipynb <==
import pandas as pd

dfs = [] 
for i in range(1, 6): 
    filename = f"./data/output/val_v14-small_fold_{i}.json"
    print(filename)
    dfs.append(pd.read_json(filename))

df = pd.concat(dfs).reset_index(drop=True)
# df = df.drop(columns=["tag_title", "checklists"])

df = df.drop(columns=["Original_Tag", "Predicted_Tag", "Domain"])

df = df.rename(columns={"tags": "Original_Tag", "predicted_tags": "Predicted_Tag"})

df.section_content.unique().shape

df = df[df.filename.str.contains('synth') == False]

print(df.shape)

df.columns

df.head()

import re

def extract_section_num(x): 
    text = x.split("\n")[1].strip().upper()
    matches = re.findall("^SECTION\s+\d+\.\d{1,2}", text)
    if matches:
        return matches[0]
    matches  = re.findall("^SECTION\s+\d+\.?\d{1,2}", text)
    if matches:
        return matches[0]
    matches  = re.findall("^\d+\.?\d{1,2}", text, flags=re.IGNORECASE)
    if matches:
        return matches[0]
    return "NOT FOUND"

df['section_number'] = df.section_content.apply(extract_section_num)

df[df.section_number == "NOT FOUND"].shape

df['match'] = df['Original_Tag'] == df["Predicted_Tag"]

df = df[['filename', 'title', 'section_number', 'section_content', 'Original_Tag', 'Predicted_Tag', "match"]]

df.shape

df.head()

df.to_excel("./data/output/predictions_v14_kfold_small.xlsx", index=False)

df.shape

df[df.Original_Tag != df.Predicted_Tag].shape
