==> tag-classificaton/4.1 InferenceDeployment.ipynb <==
# Installations and Import statements

# !pip3 install -q transformers sagemaker-inference
!pip install boto3 sagemaker -U -q
# !pip install torch==2.0.1 torchvision==0.15.2 transformers==4.31.0 -q

# !tar -czvf model.tar.gz trained_model/v9-small/* 

!rm  sources.tar.gz 
!tar -czvf sources.tar.gz ./inference.py

import boto3
from sagemaker.huggingface.model import HuggingFaceModel
import sagemaker
import json

# Upload Artifacts

s3 = boto3.client(
    's3'
)

bucket_name = 'cc-ai-bucket'
model_file = 'model.tar.gz'  
script_file = 'inference.py'
source_file = 'sources.tar.gz'

# Upload model and script
s3.upload_file(model_file, bucket_name, 'models/tag-classification/models/model.tar.gz')
s3.upload_file(script_file, bucket_name, 'models/tag-classification/scripts/inference.py')
s3.upload_file(source_file, bucket_name, 'models/tag-classification/scripts/sources.tar.gz')

# Sagemaker Role 
Needs SagemakerFullAccess policy to deploy get access to EC2 instances and deploy models

try:
    role = sagemaker.get_execution_role()
except ValueError:
    iam = boto3.client('iam')
    role = iam.get_role(RoleName='sagemaker_execution_role')['Role']['Arn']

print(f"sagemaker role arn: {role}")

sagemaker_session = sagemaker.Session(
)

# Specify Artifact Location
Location where model and source artifacts are uploaded 

bucket_name = "cc-ai-bucket/models/tag-classification"
model_artifact = f's3://{bucket_name}/models/model.tar.gz'
script_location = f's3://{bucket_name}/scripts/'
print(model_artifact)

# Deploy Hugging Face Model

# Initialize the Hugging Face model
huggingface_model = HuggingFaceModel(
    model_data=model_artifact,
    role=role,
    source_dir="s3://cc-ai-bucket/models/tag-classification/scripts/sources.tar.gz",
    entry_point="inference.py",  # Path to your custom inference script
    transformers_version="4.37.0",
    pytorch_version="2.1.0",
    py_version="py310",
    sagemaker_session=sagemaker_session
)



# Deploy the model
predictor = huggingface_model.deploy(
    initial_instance_count=1,
    instance_type="ml.c5.xlarge" # Can be changed g4 or g5 gpus 
)

print(predictor.endpoint_name)

# Test Inference

import pandas as pd

df = pd.read_json("./data/cleaned_tags_data_na_tag_lt3.json", orient='records')

sample_df = df[df.tags != "NA"].sort_values(by="word_count")
sample_df.head()

test_data = {
    "inputs": sample_df.sort_values(by="word_count", ascending=False).iloc[0].section_content
}
print(test_data)

import time 

start_time = time.time()
for t in sample_df.sample(20).itertuples(): 
    test_data = {
        "inputs": t.section_content
    }
    result = predictor.predict(json.dumps(test_data))
    result = json.loads(result)
    print(result[0], t.tags, result[0] == t.tags)
print("Time taken ", time.time() - start_time)

7.87/20

result = predictor.predict(json.dumps(test_data))
print(f"Prediction: {result}")

# Test Inference from Client 


# Initialize the SageMaker runtime client
runtime_client = boto3.client('sagemaker-runtime')

# Define the endpoint name
endpoint_name = "huggingface-pytorch-inference-2024-09-06-06-14-50-836"

# Convert input data to JSON string
payload = json.dumps(test_data)
# Invoke the endpoint
response = runtime_client.invoke_endpoint(
    EndpointName=endpoint_name,
    ContentType='application/json',
    Body=payload
)
# Parse the response
result = json.loads(response['Body'].read().decode())
print(f"Prediction: {result}")

print("Deleting Model")
predictor.delete_model()
print("Deleting Endpoint")
predictor.delete_endpoint()
